reduce cratio decimals from 18 to 4

+++ contracts/interfaces/IEscrow.sol
@@ -32,7 +32,7 @@ interface IEscrow {
 
     function borrower() external returns(address);
 
-    function minimumCollateralRatio() external returns(uint256);
+    function minimumCollateralRatio() external returns(uint32);
 
     // Functions
